the work never ends

+++ src/main/java/com/dtsx/astra/cli/gateways/db/collection/CollectionGatewayImpl.java
@@ -1,19 +1,22 @@
-package com.dtsx.astra.cli.gateways.collection;
+package com.dtsx.astra.cli.gateways.db.collection;
 
-import com.datastax.astra.client.collections.definition.CollectionDefaultIdTypes;
+import com.datastax.astra.client.collections.commands.options.ListCollectionOptions;
 import com.datastax.astra.client.collections.definition.CollectionDefinition;
 import com.datastax.astra.client.collections.definition.CollectionDescriptor;
-import com.dtsx.astra.cli.gateways.APIProvider;
-import com.dtsx.astra.cli.gateways.APIProviderImpl;
+import com.datastax.astra.client.collections.definition.documents.Document;
+import com.datastax.astra.client.core.commands.Command;
+import com.datastax.astra.client.exceptions.DataAPIException;
 import com.dtsx.astra.cli.core.models.CollectionRef;
 import com.dtsx.astra.cli.core.models.KeyspaceRef;
-import com.dtsx.astra.cli.core.exceptions.db.CollectionNotFoundException;
 import com.dtsx.astra.cli.core.output.AstraLogger;
+import com.dtsx.astra.cli.gateways.APIProvider;
+import com.dtsx.astra.cli.gateways.APIProviderImpl;
 import com.dtsx.astra.sdk.utils.AstraEnvironment;
 import lombok.RequiredArgsConstructor;
+import lombok.val;
 
 import java.util.List;
-import java.util.Map;
+import java.util.Optional;
 
 import static com.dtsx.astra.cli.core.output.AstraColors.highlight;
 
@@ -26,30 +29,32 @@ public class CollectionGatewayImpl implements CollectionGateway {
     }
 
     @Override
-    public List<String> findAllCollections(KeyspaceRef ksRef) {
+    public List<CollectionDescriptor> findAllCollections(KeyspaceRef ksRef) {
         return AstraLogger.loading("Listing collections for keyspace " + highlight(ksRef), (_) ->
-            api.dataApiDatabase(ksRef).listCollectionNames()
+            api.dataApiDatabase(ksRef).listCollections()
         );
     }
 
     @Override
-    public boolean collectionExists(CollectionRef collRef) {
+    public Optional<CollectionDefinition> findOneCollection(CollectionRef collRef) {
         try {
-            findOneCollection(collRef);
-            return true;
-        } catch (CollectionNotFoundException e) {
-            return false;
+            return AstraLogger.loading("Getting collection " + highlight(collRef), (_) -> {
+                return Optional.of(
+                    api.dataApiDatabase(collRef.keyspace()).getCollection(collRef.name()).getDefinition()
+                );
+            });
+        } catch (DataAPIException e) {
+            if (e.getErrorCode().equals("COLLECTION_NOT_EXIST")) {
+                return Optional.empty();
+            }
+            throw e;
         }
     }
 
     @Override
-    public CollectionDefinition findOneCollection(CollectionRef collRef) {
-        return AstraLogger.loading("Getting collection " + highlight(collRef), (_) -> {
-            try {
-                return api.dataApiDatabase(collRef.keyspace()).getCollection(collRef.name()).getDefinition();
-            } catch (Exception e) {
-                throw new CollectionNotFoundException(collRef);
-            }
+    public boolean collectionExists(CollectionRef collRef) {
+        return AstraLogger.loading("Checking if collection " + highlight(collRef) + " exists", (_) -> {
+            return findOneCollection(collRef).isPresent();
         });
     }
 
@@ -64,7 +69,13 @@ public class CollectionGatewayImpl implements CollectionGateway {
         String embeddingKey,
         List<String> indexingAllow,
         List<String> indexingDeny
-    ) {
+    ) throws InternalCollectionAlreadyExistsException {
+        val exists = collectionExists(collRef);
+
+        if (exists) {
+            throw new InternalCollectionAlreadyExistsException(collRef);
+        }
+
         AstraLogger.loading("Creating collection " + highlight(collRef), (_) -> {
 //            var collectionDefBuilder = new CollectionDefinition();
 //
@@ -115,14 +126,16 @@ public class CollectionGatewayImpl implements CollectionGateway {
     }
 
     @Override
-    public void deleteCollection(CollectionRef collRef) {
+    public void deleteCollection(CollectionRef collRef) throws InternalCollectionNotFoundException {
+        val exists = collectionExists(collRef);
+
+        if (!exists) {
+            throw new InternalCollectionNotFoundException(collRef);
+        }
+
         AstraLogger.loading("Deleting collection " + highlight(collRef), (_) -> {
-            try {
-                api.dataApiDatabase(collRef.keyspace()).dropCollection(collRef.name());
-                return null;
-            } catch (Exception e) {
-                throw new CollectionNotFoundException(collRef);
-            }
+            api.dataApiDatabase(collRef.keyspace()).dropCollection(collRef.name());
+            return null;
         });
     }
 }
